# UT5/Supuesto práctico proyecto/TejeraSantana_Adoney_supuesto_dsw_ut5/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Supuesto UT5 Tejera Santana Adoney",
        "description": "Documentacion de supuesto ut5",
        "version": "11.31"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Servidor principal de la API"
        }
    ],
    "paths": {
        "/api/adoneytj/ejercicios": {
            "get": {
                "tags": [
                    "Ejercicios"
                ],
                "summary": "Listar ejecicios",
                "description": "Lista todos los ejercicios, solo admin y entrenadores tienen autorización",
                "operationId": "indexEjercicios",
                "responses": {
                    "200": {
                        "description": "Operación exitosa"
                    },
                    "401": {
                        "description": "Sin autorización, debe ser admin o entrenador"
                    }
                }
            },
            "post": {
                "tags": [
                    "Ejercicios"
                ],
                "summary": "Crear un ejercicio",
                "description": "Crea un nuevo ejercicio, solo admin y entrenadores tienen autorización",
                "operationId": "storeEjercicios",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "descripcion",
                                    "id_tipo_musculo"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Sentadilla normal y corriente"
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Crecerá tu pierna"
                                    },
                                    "id_tipo_musculo": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Operación exitosa"
                    },
                    "401": {
                        "description": "Sin autorización, debe ser admin o entrenador"
                    },
                    "205": {
                        "description": "Algún parámetro inválido"
                    }
                }
            }
        },
        "/api/adoneytj/ejercicios/{id_ejercicio}": {
            "get": {
                "tags": [
                    "Ejercicios"
                ],
                "summary": "Obtener un ejecicio",
                "description": "Obtiene el ejercicio indicado, solo admin y entrenadores tienen autorización",
                "operationId": "showEjercicios",
                "parameters": [
                    {
                        "name": "id_ejercicio",
                        "in": "path",
                        "description": "ID del ejercicio",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa"
                    },
                    "401": {
                        "description": "Sin autorización, debe ser admin o entrenador"
                    }
                }
            },
            "put": {
                "tags": [
                    "Ejercicios"
                ],
                "summary": "Actualiza un ejercicio",
                "description": "Actualiza un ejercicio, solo admin y entrenadores tienen autorización",
                "operationId": "updateEjercicios",
                "parameters": [
                    {
                        "name": "id_ejercicio",
                        "in": "path",
                        "description": "ID del ejercicio",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Sentadilla normal y corriente"
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Crecerá tu pierna"
                                    },
                                    "id_tipo_musculo": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Operación exitosa"
                    },
                    "401": {
                        "description": "Sin autorización, debe ser admin o entrenador"
                    },
                    "205": {
                        "description": "Algún parámetro inválido o no existe el ejercicio"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Ejercicios"
                ],
                "summary": "Borrar un ejecicio",
                "description": "Obtiene el ejercicio indicado, solo admin y entrenadores tienen autorización",
                "operationId": "deleteEjercicios",
                "parameters": [
                    {
                        "name": "id_ejercicio",
                        "in": "path",
                        "description": "ID del ejercicio",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa"
                    },
                    "401": {
                        "description": "Sin autorización, debe ser admin o entrenador"
                    }
                }
            }
        },
...
    },
...
}